wip: add some more fortification + linting + fmt...
