+++ resources/views/tables/columns/translation-cell.blade.php
@@ -5,10 +5,7 @@
 <livewire:filament-translation-cell-editor
     :group="$state['group']"
     :translation-key="$state['translation_key']"
-    :translations="$state['translations'] ?? []"
-    :locales="$state['locales']"
-    :source-locale="$state['source_locale']"
-    :has-ai="$state['has_ai'] ?? false"
-    :ai-driver="$state['ai_driver'] ?? null"
+    :content="['translations' => $state['translations'] ?? [], 'locales' => $state['locales'], 'source_locale' => $state['source_locale']]"
+    :ai="['enabled' => $state['has_ai'] ?? false, 'driver' => $state['ai_driver'] ?? null]"
     :wire:key="$state['group'] . '.' . $state['translation_key']"
 />
